+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## [Unreleased]
+
+- The default port for the library has been changed from 15432 to 25432 to avoid clashing with instances of the CLI app.
+
 ## [0.3.0]
 
 ### Changed